Add downloader for AVCX
Adds a downloader for AVCX (American Values Club Crosswords). These
are popular crosswords from a variety of creators, see
https://avxwords.com/about-us/.

This is a subscription-only crossword series, and requires
authentication. This is handled in exactly the same way as NYT.

This downloader may not seem to serve an obvious purpose, given that
AVCX emails subscribers an AcrossLite compatible .puz file for every
new release. However, I'm thinking it will be useful for the
following features:

 * Automatic downloading of new crosswords, e.g. using a crontab.
 * Would allow downstream software like Gnome Crosswords to fetch
     AVCX automatically.
 * I do a few fixups on the .puz files and include and difficulty
     metadata.
 * Some crosswords are not available in .puz from the AVCX website,
     e.g. barred crosswords like https://avxwords.com/puzzles/1621.
     As these are available in JPZ I think it would be nice to do
     a very minimal translation into a close .puz approximation.