Fix API consistency issues
Many sub-classes modify their parents' methods and alter the allowed
arguments. This is a problem because if a function call is written on
the API of the parent class, it will result in an error if it uses
keyword arguments unavailable in the child class, and may fail or yield
unexpected results if it uses positional arguments in a different order.

Even when this can be guaranteed not to be the case through analysis of
the code, it's still a bad practice that can break newly added code
unexpectedly. I've fixed all of these issues that I could find.