Add considerably more error checking
This is mainly just catching and dealing with type mismatch errors
reported by Pyright. This catches a lot of cases where HTML parsing can
fall through, but not all of them, e.g. IndexError and KeyError are not
type errors. (But failing to prove an optional type has a value is.)

How you deal with these issues is ultimately a matter of some
preference. In two or three places, I removed a perfectly adequate check
that looked for an AttributeError, since exception catching doesn't
satisfy the type checker. But this commit could be revised to leave
these checks in or even do more of them.

A follow up to this should probably add checks on the result of every
`request.get` and `request.post`, as I only did that here where it was
convenient.