Refactor imports of downloader plugins
This commit changes how importing the downloaders happens slightly.
The point is to avoid making any assumptions in xword_dl.py about
what plugins are available and what methods they have.

This has a number of advantages, for example, that it's trivial for
someone to add a downloader plugin that requires authentication
without adding another special case to the main control flow. This
also happens to make code checkers happy as a side effect, because
they can't determine whether the plugin classes are in the downloader
namespace or not without executing the downloader __init__ code.

The work in this commit is prefatory to an additional refactor which
would create a generic Downloader class that automatically finds
plugins and abstracts the control flow away from the application logic
in xword_dl.py.