Clean up a number of small issues and improve error checking (#205)
* Fix import of urllib.parse module

Many files import urllib, but this is a package that contains the module
`urllib.parse`. This works, when it does (I haven't checked every case)
because `requests` is also imported, and it imports the module
correctly, which makes it available to xword-dl under the urllib
namespace.

* Fix reference to undefined setting

Because of the way the settings dict is populated, if the preserve_html
option is not provided, the key will not exist.

`settings.get()` usually works because `None` is false-y, but has the
potential to fail unexpectedly

* fix reference to undefined method

* Avoid use of .get() when existence is immediately assumed

Clean up some uses of get where the code either immediately assumes that
the result is not None or else provides a replacement value, e.g.
through a ternary expression.

Keeping this clean makes error messages more debuggable when unexpected
errors aren't caught.

* Fix API consistency issues

Many sub-classes modify their parents' methods and alter the allowed
arguments. This is a problem because if a function call is written on
the API of the parent class, it will result in an error if it uses
keyword arguments unavailable in the child class, and may fail or yield
unexpected results if it uses positional arguments in a different order.

Even when this can be guaranteed not to be the case through analysis of
the code, it's still a bad practice that can break newly added code
unexpectedly. I've fixed all of these issues that I could find.

* Add considerably more error checking

This is mainly just catching and dealing with type mismatch errors
reported by Pyright. This catches a lot of cases where HTML parsing can
fall through, but not all of them, e.g. IndexError and KeyError are not
type errors. (But failing to prove an optional type has a value is.)

How you deal with these issues is ultimately a matter of some
preference. In two or three places, I removed a perfectly adequate check
that looked for an AttributeError, since exception catching doesn't
satisfy the type checker. But this commit could be revised to leave
these checks in or even do more of them.

A follow up to this should probably add checks on the result of every
`request.get` and `request.post`, as I only did that here where it was
convenient.

* fix erroneous equality check used in place of an assignment

* Remove some incorrect (and probably dead) code

These lines are an implementation of `guess_date_from_id`, which is
supposed to turn a puzzle ID into a `datetime`, but they can't possibly
have worked because you can't call `datetime.strftime` that way, and in
any case the IDs now appear to be random hashes without date
information.

* handle error a bit more robustly

* remove unused imports

* avoid unnecessary 'import *' behavior

* remove unused variables and dead code

* make all exception catching explicit and refactor

* Refactor imports of downloader plugins

This commit changes how importing the downloaders happens slightly.
The point is to avoid making any assumptions in xword_dl.py about
what plugins are available and what methods they have.

This has a number of advantages, for example, that it's trivial for
someone to add a downloader plugin that requires authentication
without adding another special case to the main control flow. This
also happens to make code checkers happy as a side effect, because
they can't determine whether the plugin classes are in the downloader
namespace or not without executing the downloader __init__ code.

The work in this commit is prefatory to an additional refactor which
would create a generic Downloader class that automatically finds
plugins and abstracts the control flow away from the application logic
in xword_dl.py.

* Make import of downloader modules much cleaner

Gets rid of a lot of the complicated introspection required to pick
up the available downloader classes, and also moves all of this
logic out of xword_dl.py.

This change could potentially be amended to add the plugins to the
downloader namespace as before, but after this change along with the
last several, this has no real benefit and leaves the namespace
cluttered.

* remove unnecessary import in __init__.py

* Simplify overcomplicated use of hasattr

With this commit the entire repository is Pyright clean!!

* fix issue where refactor broke authentication

* Additional typing improvements

Simplifies __get_subclasses(), and adds type annotations to important
methods in __init__.py, xword_dl.py, and utils.py. The result of this
work was to reveal a number of issues with API inconsistency, which
this commit fixes as well.

* Make __get_subclasses generic

More appropriate as this function isn't specific to the downloader
types.

* type checking: use Python 3.11 compatible generic signature

* fix typo

* check for non-empty commands in help text

* correctly filter out downloaders that have no independent matches_url implementation

* ruff format

* update min version to 3.10 because of typing union syntax

* plugin feature detection: clean up method override checks

---------

Co-authored-by: Parker Higgins <parker@parkerhiggins.net>